pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "debate-project"
version = "0.1.0"
description = "Debate sparring agent prototype"
requires-python = ">=3.11"
dependencies = [
	"fastapi>=0.115.0,<1.0.0",
	"openai>=1.66.0,<2.0.0",
	"ddgs>=8.0.0,<10.0.0",
	"python-dotenv>=1.0.1,<2.0.0",
	"uvicorn>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
	"pytest>=8.0,<9.0",
	"pytest-cov>=5.0,<7.0",
	"ruff>=0.8,<1.0",
	"mypy>=1.13,<2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
